Add files via upload
We have been serving the industry with our expertise and people. We have been a renowned face in the industry for quality products and services. After experiencing the tedious process of tenders, quotations and orders involving extensive cost, it's time for B2B procurement. We have created a unique platform for the industry; an Online Platform i.e. www.agg-blackbox.com which will act as a bridge between the seller and the buyer. It will not only ease the process of buying and selling but also reduce the procurement lead time. It is a platform to bring a Global B2B experience.